Move server.js into lib folder
