+++ index.js
@@ -13,7 +13,7 @@ import {
   TextDocuments
 } from 'vscode-languageserver/node.js'
 import {TextDocument} from 'vscode-languageserver-textdocument'
-import {configureUnifiedLanguageServer} from './server.js'
+import {configureUnifiedLanguageServer} from './lib/server.js'
 
 /**
  * Create a language server for a unified ecosystem.
