+++ lib/server.js
@@ -5,7 +5,7 @@
  * @typedef {import('vscode-languageserver').Connection} Connection
  * @typedef {import('vscode-languageserver-textdocument').TextDocument} TextDocument
  * @typedef {import('vscode-languageserver').TextDocuments<TextDocument>} TextDocuments
- * @typedef {import('.').Options} Options
+ * @typedef {import('..').Options} Options
  */
 
 import {PassThrough} from 'node:stream'
