+++ package.json
@@ -16,7 +16,7 @@
   "files": [
     "index.js",
     "index.d.ts",
-    "server.js"
+    "lib/*.js"
   ],
   "scripts": {
     "build": "rimraf '*.d.ts' 'test/*.d.ts' && tsc && type-coverage",
